+++ lib/core/src/control/mod.rs
@@ -1,0 +1,2 @@
+pub mod callback;
+pub mod event_loop;
