+++ lib/core/src/data/mod.rs
@@ -1,3 +1,4 @@
+pub mod container;
 pub mod function;
-pub mod seq;
 pub mod opt_vec;
+pub mod seq;
