+++ lib/core/src/lib.rs
@@ -24,6 +24,7 @@
 // === Module Structure Reexport ===
 // =================================
 
+pub mod control;
 pub mod data;
 pub mod dirty;
 pub mod display;
